requirements.txt
pytesseract
Pillow
pdf2image

spacy
https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
nltk

cryptography
bcrypt

Flask
Werkzeug

sqlalchemy
alembic

PyPDF2
reportlab
python-magic

google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2

click
rich

python-dotenv
tqdm

pytest
pytest-cov
pytest-mock

loguru
